+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.0.7 - Jun 7, 2025
+
+### Added
+
+- Add `useBuilding` function in `building.ts`.
+- Add `updateImageSrcByAlt` function in `building.ts`.
+- Add `SA` in goodsUrlByEra in `constants.ts`.
+
 ## v1.0.6 - May 18, 2025
 
 ### Added